+++ examples/default_priors/kima_setup.cpp
@@ -1,9 +1,7 @@
-#include <iostream>
 #include "DNest4.h"
 #include "Data.h"
 #include "RVmodel.h"
 
-using namespace std;
 using namespace DNest4;
 
 /* priors */
@@ -16,10 +14,7 @@ const bool hyperpriors = false;
 const bool trend = false;
 
 // options for the model
-RVmodel::RVmodel()
-    :planets(5, 1, true, RVConditionalPrior())
-    ,mu(Data::get_instance().N())
-    ,C(Data::get_instance().N(), Data::get_instance().N())
+RVmodel::RVmodel():fix(true),npmax(1)
 {}